test(angular): prevent vite from bundle root injector singleton twice (#38)
Angular adds `@testronaut/angular/browser` to vite's `optimizeDeps.include` which somehow forces it
to be prebundled but vite bundles it with `@angular/core` so we end up with two instances of
`@angular/core` thus causing https://angular.dev/errors/NG0203 error

This happens because the test server is not importing anything other than `@testronaut/angular`.

`prebundle: {exclude...}` is a good solution but it's user-land. We'd love to find better.