fix(angular,core): remove commonjs bundles (#58)
Node.js 22.12.0 and higher support `require(esm)`.

We thought that Playwright was always transforming the esm files thus causing `exports`
error when transforming `import.meta.url` but this seems to only happen in our workspace
because the packages in `node_modules` are aliases that resolve to `packages/*/dist` so
Playwright sees them as source code to transform.